app/requirements/base.txt
Django==3.2
psycopg2-binary==2.8.5
celery==4.4.1
redis==3.4.1
django-post-office==3.5.3
phonenumbers==8.12.24
django-phonenumber-field[phonenumbers]==5.0.0
django-widget-tweaks==1.4.8
six==1.15.0
django-tailwind==2.0.1
djangorestframework==3.12.4
bleach==3.3.0
Pillow==8.2.0